requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens @ file:///C:/Users/task_177313475117182/croot/asttokens_1773135663340/work
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle @ file:///C:/miniconda3/conda-bld/cloudpickle_1764930556434/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm @ file:///C:/miniconda3/conda-bld/comm_1763119233835/work
cryptography==46.0.5
debugpy @ file:///C:/miniconda3/conda-bld/debugpy_1762421657095/work
decorator @ file:///C:/miniconda3/conda-bld/decorator_1757341256088/work
distro==1.9.0
exceptiongroup @ file:///C:/miniconda3/conda-bld/exceptiongroup_1761560437674/work
executing @ file:///C:/miniconda3/conda-bld/executing_1757061252959/work
fastapi==0.135.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.67.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ipykernel @ file:///C:/miniconda3/conda-bld/ipykernel_1762946093903/work
ipython @ file:///C:/b/abs_8eyhzleyrk/croot/ipython_1734548134403/work
jedi @ file:///C:/b/abs_3a2kbnlclc/croot/jedi_1733987412687/work
jupyter_client @ file:///C:/Users/task_176839864521852/croot/jupyter_client_1768399093953/work
jupyter_core @ file:///C:/miniconda3/conda-bld/jupyter_core_1764588347534/work
matplotlib-inline @ file:///C:/miniconda3/conda-bld/matplotlib-inline_1762779265381/work
nest-asyncio @ file:///C:/b/abs_65d6lblmoi/croot/nest-asyncio_1708532721305/work
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
parso @ file:///C:/miniconda3/conda-bld/parso_1762781925562/work
platformdirs @ file:///C:/miniconda3/conda-bld/platformdirs_1762356608986/work
prompt_toolkit @ file:///C:/miniconda3/conda-bld/prompt-toolkit_1761745021230/work
proto-plus==1.27.1
protobuf==5.29.6
psutil @ file:///C:/miniconda3/conda-bld/psutil_1761896539676/work
pure_eval @ file:///C:/miniconda3/conda-bld/pure_eval_1757067068391/work
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments @ file:///C:/miniconda3/conda-bld/pygments_1762431425987/work
pyparsing==3.3.2
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
pywin32 @ file:///C:/b/abs_6c9w_vp9zi/croot/pywin32_1754670534373/work
pyzmq @ file:///C:/miniconda3/conda-bld/pyzmq_1762375610449/work
requests==2.32.5
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
sniffio==1.3.1
spyder-kernels @ file:///C:/b/abs_e5u6y4ldr2/croot/spyder-kernels_1707937767956/work
stack_data @ file:///C:/miniconda3/conda-bld/stack_data_1757067051465/work
starlette==0.52.1
tenacity==9.1.4
tornado @ file:///C:/Users/task_176599267555405/croot/tornado_1765992779662/work
tqdm==4.67.3
traitlets @ file:///C:/b/abs_bfsnoxl4pq/croot/traitlets_1718227069245/work
typing-inspection==0.4.2
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
wcwidth @ file:///C:/Users/task_176777900206875/croot/wcwidth_1767779217517/work
websockets==16.0
